+++ src/classes/TssSecretLookup.class.ps1
@@ -1,0 +1,6 @@
+﻿class TssSecretLookup {
+    [int]$SecretId
+    [int]$FolderId
+    [int]$SecretTemplateId
+    [string]$SecretName
+}