+++ tests/functions/Find-TssSecret.Tests.ps1
@@ -1,0 +1,51 @@
+BeforeDiscovery {
+    $commandName = Split-Path ($PSCommandPath.Replace('.Tests.ps1','')) -Leaf
+    . ([IO.Path]::Combine([string]$PSScriptRoot, '..', 'constants.ps1'))
+}
+Describe "$commandName verify parameters" {
+    BeforeDiscovery {
+        [object[]]$knownParameters = 'TssSession','Raw','Permission','Scope',
+        # Folder param set
+        'FolderId','IncludeSubFolders',
+        # field param set
+        'Field','FieldText','ExactMatch','FieldSlug','ExtendedField','ExtendedTypeId',
+        # secret param set
+        'SecretTemplateId','SiteId','HeartbeatStatus','IncludeInactive','ExcludeActive','RpcEnabled','SharedWithMe','PasswordTypeIds','ExcludeDoubleLock','DoubleLockId'
+        [object[]]$currentParams = ([Management.Automation.CommandMetaData]$ExecutionContext.SessionState.InvokeCommand.GetCommand($commandName,'Function')).Parameters.Keys
+        [object[]]$commandDetails = [System.Management.Automation.CommandInfo]$ExecutionContext.SessionState.InvokeCommand.GetCommand($commandName,'Function')
+        $unknownParameters = Compare-Object -ReferenceObject $knownParameters -DifferenceObject $currentParams -PassThru
+    }
+    Context "Verify parmaeters" -Foreach @{currentParams = $currentParams} {
+        It "$commandName should contain <_> parameter" -TestCases $knownParameters {
+            $_ -in $currentParams | Should -Be $true
+        }
+        It "$commandName should not contain parameter: <_>" -TestCases $unknownParameters {
+            $_ | Should -BeNullOrEmpty
+        }
+    }
+    Context "Command specific details" {
+        It "$commandName should set OutputType to TssSecretLookup" -TestCases $commandDetails {
+            $_.OutputType.Name | Should -Be 'TssSecretLookup'
+        }
+    }
+}
+Describe "$commandName works" {
+    BeforeDiscovery {
+        $session = New-TssSession -SecretServer $ss -Credential $ssCred
+        $object = Find-TssSecret $session -RpcEnabled -SecretTemplateId 6003
+        $session.SessionExpire()
+
+        $props = 'SecretId','FolderId','SecretTemplateId','SecretName'
+    }
+    Context "Checking" -Foreach @{object = $object} {
+        It "Should not be empty" {
+            $object | Should -Not -BeNullOrEmpty
+        }
+        It "Should find one secret with RPC enabled" {
+            $object.SecretTemplateId | Should -Be 6003
+        }
+        It "Should output <_> property" -TestCases $props {
+            $object.PSObject.Properties.Name | Should -Contain $_
+        }
+    }
+}